Add files via upload
This is just a simple python script to delete a bunch of .srt files in a folder. Very simple but comes in handy because I don't need them for anything and a lot is a pain to delete. So this is just for those who searched for this and don't know Python. Most everyone else cane write one themselves. Maybe even better.